requirements.txt
streamlit>=1.32.0
pandas>=2.2.0
notion-client>=2.0.0
python-dotenv>=1.0.0
streamlit-mermaid>=0.0.3
langchain>=0.1.0
langchain-google-genai>=0.0.5
langgraph>=0.0.10 